% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp_misto.R
\name{ahp_geral}
\alias{ahp_geral}
\title{ahp_geral}
\usage{
ahp_geral(objeto, mapeamento = "PADRAO", nomes_alternativas = "PADRAO")
}
\arguments{
\item{objeto}{Lista de matrizes pareadas ou caminho do excel contendo as
matrizes pareadas devidamente formatadas.}

\item{mapeamento}{Vetor contendo a quantidade de subcriterios de cada criterio
da sua hierarquia, da esquerda para direita. Se nao preenchido o padrao
e preencher com 0. Em caso de duvida, veja a vignette tutorial.}

\item{nome_alternativas}{Vetor contendo os nomes das alternativas da sua hierarquia,
se nao preenchido retorna um vetor de LETTERS[1:qtdAlternativas]}
}
\value{
Tabela contendo as relacoes dos criterios, subcriterios (Se houver) e Alternativas
Utilizando o metodo AHP.
}
\description{
Realiza o calculo da AHP em uma lista de matrizes pareadas ou em um
endereco de planilhas do excel devidamente formatados.
}
