% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_funcoes.R
\name{aplica_ahp}
\alias{aplica_ahp}
\title{AHP excel}
\usage{
aplica_ahp(caminho, alternativas = "PADRAO", cores = "PADRAO")
}
\arguments{
\item{caminho}{e uma string do caminho do arquivo excel ou uma lista com as matrizes}

\item{alternativas}{espera um vetor com os nomes das alternativas a serem estudadas.
Se "PADRAO" ira preencher o nome das alternativas com letras de A:Z}

\item{cores}{altera as cores do fundo das celulas da tabela. pode ser usados
"PADRAO", "CINZA", "BRANCO"}
}
\value{
tabela formatada para visualizacao
}
\description{
Aplica AHP em um grupo de matrizes devidamente formatadas em um arquivo excel.
}
