% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentacao_bases.R
\docType{data}
\name{BD1}
\alias{BD1}
\title{BD1}
\format{
Lista com 6 matrizes de julgamento

[1] Matriz para comparacao dos criterios.

[2] Matriz para comparacao das alternativas em luz do  criterio Conforto (CF).

[3] Matriz para comparacao das alternativas em luz do criterio Aquisicao (AQ).

[4] Matriz para comparacao das alternativas em luz do criterio Prestigio (PS).

[5] Matriz para comparacao das alternativas em luz do criterio Revenda (RV).

[6] Matriz para comparacao das alternativas em luz do criterio Manutencao (MA).
}
\source{
BD1
}
\usage{
BD1
}
\description{
Lista de matrizes de julgamento simples (sem subcriterios) de um exemplo para escolha de um carro com 3 opcoes de alternativas.
Os criterios adotados foram Conforto,
aquisicao, prestigio, revenda e manutencao.
}
\keyword{datasets}
