% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentacao_bases.R
\docType{data}
\name{BD2}
\alias{BD2}
\title{BD2}
\format{
Lista com 5 matrizes de julgamento

[1] Matriz para comparacao dos criterios.

[2] Matriz para comparacao das alternativas em luz do  criterio Custo.

[3] Matriz para comparacao das alternativas em luz do criterio Paisagem.

[4] Matriz para comparacao das alternativas em luz do criterio Prestigio.

[5] Matriz para comparacao das alternativas em luz do criterio Lazer.
}
\source{
BD2
}
\usage{
BD2
}
\description{
Lista de matrizes de julgamento simples (sem subcriterios) de um exemplo para escolha de um lugar para viajar com 3 opcoes de alternativas.
Os criterios adotados foram custo,
paisagem, prestigio e lazer.
}
\keyword{datasets}
