% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentacao_bases.R
\docType{data}
\name{BD3}
\alias{BD3}
\title{BD3}
\format{Lista com 13 matrizes de julgamento

[1] Matriz para comparacao dos criterios.

[2] Matriz para comparacao dos subcriterios do criterio financeiro.

[3] Matriz para comparacao dos subcriterios do criterio estrategico.

[4] Matriz para comparacao dos subcriterios do criterio tecnologico.

[5] Matriz para comparacao das alternativas em luz do subcriterio tempo de retorno.

[6] Matriz para comparacao das alternativas em luz do subcriterio indice de lucratividade.

[7] Matriz para comparacao das alternativas em luz do subcriterio valor presente.

[8] Matriz para comparacao das alternativas em luz do subcriterio melhorar processos.

[9] Matriz para comparacao das alternativas em luz do subcriterio reputacao.

[10] Matriz para comparacao das alternativas em luz do subcriterio riscos.

[11] Matriz para comparacao das alternativas em luz do subcriterio capacidade do time.

[12] Matriz para comparacao das alternativas em luz do subcriterio compatibilidade.

[13] Matriz para comparacao das alternativas em luz do subcriterio custo total de propriedade.}
\source{
BD3
}
\usage{
BD3
}
\description{
Lista de matrizes de julgamentos com subcriterios para um exemplo de escolha de um plano de priorizacao
de servicos para uma empresa. Os criterios adotados sao: financeiro com subcriterios tempo de retorno, indice
de lucratividade e Valor presente; Estrategico com subcriterios melhorar processos, melhorar reputacao e
reducao de riscos; Tecnologico com subcriterios capacidade do time, compatibilidade e custo total de prioridade.
Para utilizacao na funcao ahp_geral, o argumento mapeamento deve ser preenchido como c(3,3,3), veja a vignette
tutorial caso haja duvida.
}
\keyword{datasets}
