% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_funcoes.R
\name{calcula_prioridades}
\alias{calcula_prioridades}
\title{calcula_prioridades}
\usage{
calcula_prioridades(lista)
}
\arguments{
\item{lista}{Lista de matrizes pareadas}
}
\value{
Retorna uma lista contendo vetores de prioridades para cada matriz na lista lida
}
\description{
Calcula o vetor de prioridades de uma matriz pareada
}
