% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_funcoes.R
\name{CR}
\alias{CR}
\title{CR}
\usage{
CR(matriz)
}
\arguments{
\item{Matriz}{pareada}
}
\value{
Taxa de consistencia de saaty
}
\description{
Calcula a taxa de consistencia de saaty para uma matriz pareada
}
