% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tabela.R
\name{formata_tabela}
\alias{formata_tabela}
\title{formata_tabela}
\usage{
formata_tabela(tabela, cores = "PADRAO")
}
\arguments{
\item{tabela}{Tabela AHP criada pela funcao ahp_geral()}

\item{cores}{Padrao de cores para formatar a tabela. Se "PADRAO" retorna o padrao de cores (verde, azul, verde ou azul); se "CINZA" retorna o padrao de cores de cinza;
se "BRANCO" retorna a tabela sem cores}
}
\value{
Retorna uma tabela formatada com cores defundo responsivas as quantidades de prioridade dos elementos
}
\description{
Formata uma tabela AHP criada pela funcao ahp_geral()
}
