% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_funcoes.R
\name{ler}
\alias{ler}
\title{ler}
\usage{
ler(caminho)
}
\arguments{
\item{caminho}{Endereco para um arquivo excel que contem as planilhas}
}
\value{
Retorna uma lista contendo as matrizes pareadas do arquivo excel
}
\description{
Le um arquivo do excel contendo as matrizes pareadas e transforma todas suas planilhas em uma lista de matrizes no R
}
