% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_funcoes.R
\name{matriz_julgamento}
\alias{matriz_julgamento}
\title{matriz_julgamento}
\usage{
matriz_julgamento(qtd_comparacoes, CR = TRUE, qtd_matrizes = 1)
}
\arguments{
\item{qtd_comparacoes}{Quantida de elementos para serem avaliados}

\item{CR}{Se TRUE retorna tambem a taxa de consistencia de saaty, se FALSE retorna apenas matriz}

\item{qtd_matrizes}{Quantidade de matrizes a serem criadas}
}
\value{
Retorna uma lista com 2 posicoes. Primeira posicao contem as matrizes pareadas e a segunda posicao as suas taxas de consistencia
}
\description{
Cria matrizes pareadas e pode testar taxa de consistencia de saaty
}
