% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp_misto.R
\name{ranque}
\alias{ranque}
\title{Ranque}
\usage{
ranque(tabela)
}
\arguments{
\item{tabela}{tabela AHP criada pela cuncao ahp_geral()}
}
\value{
Ranque das alternativas em ordem crescente
}
\description{
Cria o ranque de alternativas em ordem crescente de uma tabela AHP
criada pela funcao ahp_geral()
}
